pulling attack updates

See merge request cpts487-sp2021/tbd!48

+++ BulletHell/BulletHell/Game Utilities/GameLoader.cs
@@ -55,7 +55,7 @@
 
             if (o is long)
             {
-                return Convert.ToInt32(o);
+                return Convert.ToDouble(o);
             }
 
             return o;
